Improved: code as per the updations in the UI to include the side menus as components(#1wut90a)
